+++ src/main/java/madsciencemod/patches/CardTagsEnum.java
@@ -1,0 +1,9 @@
+package madsciencemod.patches;
+
+import com.evacipated.cardcrawl.modthespire.lib.SpireEnum;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+
+public class CardTagsEnum {
+    @SpireEnum
+    public static AbstractCard.CardTags TRINKET_TAG;
+}