+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.3.1
 * Fix: FuelPotion.makeCopy() now returns a FuelPotion
+* Fix: Weaponized trinkets now count as trinkets for Minature Weapons and Gather Parts.
 
 ### v0.3.0
 * Prefix all cards, relics, powers and potions with ModID